+++ packages/terra-graphs/tests/wdio/Graphs/PairedResultGraph/paired-result-graph-spec.js
@@ -83,5 +83,101 @@ Terra.describeViewports('PairedResultGraph ', ['tiny', 'medium', 'large'], () =>
         Terra.validates.screenshot('multiple_region_hovered_on_second_data_low', { selector: '.carbon-graph-container' });
       });
     });
+
+    describe('Reflow', () => {
+      let panningBtn = null;
+      describe('Dynamic data', () => {
+        beforeEach(() => {
+          browser.url('/raw/tests/cerner-terra-graphs-docs/graphs/paired-result/reflow/dynamic-data');
+        });
+        it('validates panning left', () => {
+          panningBtn = $('.button-pan-left');
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_dynamic_data_panning_left_1', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_dynamic_data_panning_left_2', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_dynamic_data_panning_left_3', { selector: '.carbon-graph-container' });
+        });
+        it('validates panning right', () => {
+          panningBtn = $('.button-pan-right');
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_dynamic_data_panning_right_1', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_dynamic_data_panning_right_2', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_dynamic_data_panning_right_3', { selector: '.carbon-graph-container' });
+        });
+      });
+      describe('Dynamic eventline', () => {
+        beforeEach(() => {
+          browser.url('/raw/tests/cerner-terra-graphs-docs/graphs/paired-result/reflow/dynamic-eventline');
+        });
+        it('validates panning left', () => {
+          panningBtn = $('.button-pan-left');
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_dynamic_eventline_panning_left_1', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_dynamic_eventline_panning_left_2', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_dynamic_eventline_panning_left_3', { selector: '.carbon-graph-container' });
+        });
+        it('validates panning right', () => {
+          panningBtn = $('.button-pan-right');
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_dynamic_eventline_panning_right_1', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_dynamic_eventline_panning_right_2', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_dynamic_eventline_panning_right_3', { selector: '.carbon-graph-container' });
+        });
+      });
+      describe('Multiple Datasets', () => {
+        beforeEach(() => {
+          browser.url('/raw/tests/cerner-terra-graphs-docs/graphs/paired-result/reflow/multiple-datasets-with-y-2-axis');
+        });
+        it('validates panning left', () => {
+          panningBtn = $('.button-pan-left');
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_multiple_datasets_panning_left_1', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_multiple_datasets_panning_left_2', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_multiple_datasets_panning_left_3', { selector: '.carbon-graph-container' });
+        });
+        it('validates panning right', () => {
+          panningBtn = $('.button-pan-right');
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_multiple_datasets_panning_right_1', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_multiple_datasets_panning_right_2', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_multiple_datasets_panning_right_3', { selector: '.carbon-graph-container' });
+        });
+      });
+      describe('Simple', () => {
+        beforeEach(() => {
+          browser.url('/raw/tests/cerner-terra-graphs-docs/graphs/paired-result/reflow/simple-panning');
+        });
+        it('validates panning left', () => {
+          panningBtn = $('.button-pan-left');
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_simple_panning_left_1', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_simple_panning_left_2', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_simple_panning_left_3', { selector: '.carbon-graph-container' });
+        });
+        it('validates panning right', () => {
+          panningBtn = $('.button-pan-right');
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_simple_panning_right_1', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_simple_panning_right_2', { selector: '.carbon-graph-container' });
+          panningBtn.click();
+          Terra.validates.screenshot('paired_result_simple_panning_right_3', { selector: '.carbon-graph-container' });
+        });
+      });
+    });
   });
 });