+++ packages/terra-graphs-docs/CHANGELOG.md
@@ -13,3 +13,4 @@
   * Added data objects and examples for Combination graph.
   * Added reflow examples for Line Graph.
   * Added critical and basic example for timeline graph.
+  * Added reflow examples for Paired result graph.